chore: refactor createPostmanSidecar function
