+++ cmd/internal/kube/inject.go
@@ -7,7 +7,6 @@ import (
 	"github.com/akitasoftware/akita-libs/akid"
 	"github.com/akitasoftware/go-utils/optionals"
 	"github.com/pkg/errors"
-	"github.com/postmanlabs/postman-insights-agent/cfg"
 	"github.com/postmanlabs/postman-insights-agent/cmd/internal/cmderr"
 	"github.com/postmanlabs/postman-insights-agent/cmd/internal/kube/injector"
 	"github.com/postmanlabs/postman-insights-agent/printer"
@@ -160,70 +159,6 @@ type secretGenerationOptions struct {
 	ShouldInject bool
 	// The path to the secret file
 	Filepath optionals.Optional[string]
-}
-
-// The image to use for the Postman Insights Agent sidecar
-const akitaImage = "docker.postman.com/postman-insights-agent:latest"
-
-func createPostmanSidecar(insightsProjectID string, addAPIKeyAsSecret bool) v1.Container {
-	args := []string{"apidump", "--project", insightsProjectID}
-
-	// If a nondefault --domain flag was used, specify it for the container as well.
-	if rest.Domain != rest.DefaultDomain() {
-		args = append(args, "--domain", rest.Domain)
-	}
-
-	pmKey, pmEnv := cfg.GetPostmanAPIKeyAndEnvironment()
-	envs := []v1.EnvVar{}
-
-	if addAPIKeyAsSecret {
-		envs = append(envs, v1.EnvVar{
-			Name: "POSTMAN_API_KEY",
-			ValueFrom: &v1.EnvVarSource{
-				SecretKeyRef: &v1.SecretKeySelector{
-					LocalObjectReference: v1.LocalObjectReference{
-						Name: "postman-agent-secrets",
-					},
-					Key: "postman-api-key",
-				},
-			},
-		})
-	} else {
-		envs = append(envs, v1.EnvVar{
-			Name:  "POSTMAN_API_KEY",
-			Value: pmKey,
-		})
-	}
-
-	if pmEnv != "" {
-		envs = append(envs, v1.EnvVar{
-			Name:  "POSTMAN_ENV",
-			Value: pmEnv,
-		})
-	}
-
-	sidecar := v1.Container{
-		Name:  "postman-insights-agent",
-		Image: akitaImage,
-		Env:   envs,
-		Lifecycle: &v1.Lifecycle{
-			PreStop: &v1.LifecycleHandler{
-				Exec: &v1.ExecAction{
-					Command: []string{
-						"/bin/sh",
-						"-c",
-						"POSTMAN_INSIGHTS_AGENT_PID=$(pgrep postman-insights-agent) && kill -2 $POSTMAN_INSIGHTS_AGENT_PID && tail -f /proc/$POSTMAN_INSIGHTS_AGENT_PID/fd/1",
-					},
-				},
-			},
-		},
-		Args: args,
-		SecurityContext: &v1.SecurityContext{
-			Capabilities: &v1.Capabilities{Add: []v1.Capability{"NET_RAW"}},
-		},
-	}
-
-	return sidecar
 }
 
 // Parses the given value for the `--secret` option.
